Added testLastInDiscard() and testActionsChange()
These functions will be used to ensure the functionality of the tested card conforms to the specs (i.e. ensuring the last card in the discard pile is the card played and that the number of actions change appropriately, respectively). None of this has been tested by a compiler, so pull the changes and test on home machine ASAP.